# docs/openapi-3.0.0.yaml
openapi: 3.0.0
info:
  title: 文件闪传系统 API
  description: 一个基于FastAPI的文件闪传系统，支持端到端加密、取件码分享、大文件分片传输等功能
  version: 1.0.0
  contact:
    name: 开发团队
    email: contact@example.com
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
  - url: https://api.example.com
    description: 生产服务器
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 使用Bearer令牌进行认证，令牌由登录接口返回
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/register:
    post:
      tags:
        - 认证管理
      summary: 用户注册
      description: 创建新用户并返回访问令牌
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: 注册失败，用户名已存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/login:
    post:
      tags:
        - 认证管理
      summary: 用户登录
      description: 验证用户凭据并返回访问令牌
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: 登录失败，用户名或密码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/verify:
    get:
      tags:
        - 认证管理
      summary: 验证访问令牌
      description: 验证访问令牌的有效性并返回用户信息
      operationId: verifyToken
      responses:
        '200':
          description: 令牌有效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyTokenResponse'
        '400':
          description: 令牌无效或已过期
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # 取件码管理
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 400
        msg:
          type: string
          example: 操作失败
        data:
          type: object
      required:
        - code
        - msg

    # 认证相关模型
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          description: 用户名
          example: user123
        password:
          type: string
          minLength: 6
          maxLength: 64
          description: 密码（前端已哈希，SHA-256，64字符）
          example: e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855
      required:
        - username
        - password
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          description: 用户名
          example: user123
        password:
          type: string
          description: 密码
          example: e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855
      required:
        - username
        - password
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: user123
        created_at:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00Z
      required:
        - id
        - username
        - created_at
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        token_type:
          type: string
          example: bearer
        user:
          $ref: '#/components/schemas/UserResponse'
      required:
        - access_token
        - token_type
        - user
    VerifyTokenResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserResponse'
      required:
        - user

    # 取件码相关模型
...